chore: move handler implementations into a relay.go file.
